Fixed a warning in the backtrace() function
On some systems it returns size_t instead of int.

+++ src/common/nullpo.c
@@ -43,11 +43,11 @@ void assert_report(const char *file, int line, const char *func, const char *tar
 	ShowError("--- %s --------------------------------------------\n", title);
 	ShowError("%s:%d: '%s' in function `%s'\n", file, line, targetname, func);
 #ifdef __GNUC__
-	size = backtrace (array, 10);
-	strings = backtrace_symbols (array, size);
+	size = (int)backtrace(array, 10);
+	strings = backtrace_symbols(array, size);
 	for (i = 0; i < size; i++)
 		ShowError("%s\n", strings[i]);
-	free (strings);
+	free(strings);
 #endif
 	ShowError("--- end %s ----------------------------------------\n", title);
 }
